requirement.txt
python-dotenv
openai
google-genai
openpyxl
chromadb
python-docx